requirements.txt
Django==2.2.5
django-filter==2.2.0
django-jenkins==0.110.0
djangorestframework==3.10.3
Markdown==3.1.1
pytz==2019.2
selenium==3.141.0
sqlparse==0.3.0
urllib3==1.25.3
